Change uint8 data assumption in dataset import code

+++ training/dataset.py
@@ -38,7 +38,7 @@ class TFRecordDataset:
         self.resolution         = None
         self.resolution_log2    = None
         self.shape              = []        # [channels, height, width]
-        self.dtype              = 'uint8'
+        self.dtype              = 'float32'
         self.label_file         = label_file
         self.label_size         = None      # components
         self.label_dtype        = None
@@ -207,7 +207,7 @@ class TFRecordDataset:
         features = tf.parse_single_example(record, features={
             'shape': tf.FixedLenFeature([3], tf.int64),
             'data': tf.FixedLenFeature([], tf.string)})
-        data = tf.decode_raw(features['data'], tf.uint8)
+        data = tf.decode_raw(features['data'], tf.float32)
         return tf.reshape(data, features['shape'])
 
     # Parse individual image from a tfrecords file into NumPy array.
@@ -217,7 +217,7 @@ class TFRecordDataset:
         ex.ParseFromString(record)
         shape = ex.features.feature['shape'].int64_list.value # pylint: disable=no-member
         data = ex.features.feature['data'].bytes_list.value[0] # pylint: disable=no-member
-        return np.fromstring(data, np.uint8).reshape(shape)
+        return np.fromstring(data, np.float32).reshape(shape)
 
 #----------------------------------------------------------------------------
 # Construct a dataset object using the given options.
